Use OnDemandCompactionScheduler instead of fail_point in compactor
