+++ slatedb/src/db/builder.rs
@@ -477,7 +477,6 @@ impl<P: Into<Path>> DbBuilder<P> {
                 inner.stat_registry.clone(),
                 system_clock.clone(),
                 self.cancellation_token.clone(),
-                self.fp_registry.clone(),
             );
             let this_compactor = compactor.clone();
             // Spawn the compactor on the compaction runtime
@@ -667,7 +666,6 @@ pub struct CompactorBuilder<P: Into<Path>> {
     stat_registry: Arc<StatRegistry>,
     cancellation_token: CancellationToken,
     system_clock: Arc<dyn SystemClock>,
-    fp_registry: Arc<FailPointRegistry>,
 }
 
 #[allow(unused)]
@@ -682,7 +680,6 @@ impl<P: Into<Path>> CompactorBuilder<P> {
             stat_registry: Arc::new(StatRegistry::new()),
             cancellation_token: CancellationToken::new(),
             system_clock: Arc::new(DefaultSystemClock::default()),
-            fp_registry: Arc::new(FailPointRegistry::new()),
         }
     }
 
@@ -718,12 +715,6 @@ impl<P: Into<Path>> CompactorBuilder<P> {
         self
     }
 
-    /// Sets the fail point registry to use for the compactor.
-    pub fn with_fp_registry(mut self, fp_registry: Arc<FailPointRegistry>) -> Self {
-        self.fp_registry = fp_registry;
-        self
-    }
-
     /// Builds and returns a Compactor instance.
     pub fn build(self) -> Compactor {
         let path: Path = self.path.into();
@@ -742,7 +733,6 @@ impl<P: Into<Path>> CompactorBuilder<P> {
             self.stat_registry,
             self.system_clock,
             self.cancellation_token,
-            self.fp_registry,
         )
     }
 }
